Add uart_two_proc.vhd
This implements the UART transmit state machine using separate processes
for the combinatorial elements and the sequential elements.

I'm using the naming convention of appending "_r" on sequential elements
(registers) and appending "_s" on combinatorial elements (signals).

The port interface is slightly changed, specifically that the signal
tx_data_i is no longer reversed. So to send the letter "N", simply set
tx_data_i to the value X"4E".

There is therefore no more need for the BITSHIFT function in
toplevel.vhd